+++ dandy/agent/resource.py
@@ -8,12 +8,15 @@ from dandy.intel import BaseIntel
 
 
 class BaseAgentResource(ABC):
-    def __init__(self, processor: Type[BaseProcessor]):
+    def __init__(
+            self,
+            processor: Type[BaseProcessor]
+    ):
         if not issubclass(processor, BaseProcessor):
             raise AgentCriticalException(f'{processor} is not a sub class of "BaseProcessor"')
 
         self.processor = processor
 
     @abstractmethod
-    def use(self, *args, **kwargs) -> BaseIntel:
+    def use(self, *args, **kwargs) -> BaseIntel | None:
         raise NotImplementedError