+++ dandy/llm/agent/llm_plan.py
@@ -1,8 +1,15 @@
+from pydantic import Field
+from typing_extensions import List
+
 from dandy.agent.plan.plan import AgentPlanIntel
+from dandy.llm.agent.llm_task import LlmAgentTaskIntel
 from dandy.llm.prompt.prompt import Prompt
 
 
 class LlmAgentPlanIntel(AgentPlanIntel):
+    tasks: List[LlmAgentTaskIntel] = Field(default_factory=list)
+
+
     def to_prompt(self) -> Prompt:
         prompt = Prompt()