+++ dandy/llm/agent/llm_strategy.py
@@ -2,7 +2,9 @@ from abc import ABC
 
 from typing_extensions import List, Iterable, Type
 
+from dandy.agent.resource import BaseAgentResource
 from dandy.agent.strategy import BaseAgentStrategy
+from dandy.llm.agent.llm_resource import LlmAgentResource
 from dandy.llm.bot.llm_bot import LlmBot
 from dandy.llm.map.llm_map import BaseLlmMap
 
@@ -14,6 +16,7 @@ class BaseLlmAgentStrategy(BaseAgentStrategy, ABC):
         'workflows',
         'maps',
     ]
+    _resource_class: Type[BaseAgentResource] = LlmAgentResource
 
     maps: Iterable[Type[BaseLlmMap]] | None = None